Updated file structure,tests, pylint and github workflow
