+++ .github/workflows/workflow.yml
@@ -15,11 +15,17 @@ jobs:
       uses: actions/setup-python@master
       with:
         python-version: 3.11
-    - name: Generate coverage report
+    - name: Install requirements
       run: |
         pip install -r requirements_dev.txt
-        pip install pytest
-        pip install pytest-cov
+    - name: Run Pylint
+      run: |
+        pylint --fail-under=10 quote_gen
+    - name: Run flake8
+      run: |
+        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+    - name: Generate coverage report
+      run: |
         pytest --cov=./ --cov-report=xml
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v3
