+++ quote_gen/interface/__init__.py
