+++ quote_gen/resources/quotes/mcgregor_quotes.csv
