+++ quote_gen/resources/quotes/musk_quotes.csv
