+++ requirements_dev.txt
@@ -2,8 +2,10 @@ anyio==3.6.2
 astroid==2.12.13
 attrs==22.2.0
 click==8.1.3
+coverage==7.0.4
 dill==0.3.6
 fastapi==0.88.0
+flake8==6.0.0
 h11==0.14.0
 idna==3.4
 iniconfig==1.1.1
@@ -15,9 +17,12 @@ packaging==22.0
 pandas==1.5.2
 platformdirs==2.6.2
 pluggy==1.0.0
+pycodestyle==2.10.0
 pydantic==1.10.4
+pyflakes==3.0.1
 pylint==2.15.9
 pytest==7.2.0
+pytest-cov==4.0.0
 python-dateutil==2.8.2
 pytz==2022.7
 .
