+++ setup.cfg
@@ -16,5 +16,3 @@ universal = 1
 
 [flake8]
 exclude = docs
-[tool:pytest]
-collect_ignore = ['setup.py']
